Refactored multi-modal parser, added extra config to support it
