+++ backend/modules/parsers/__init__.py
@@ -1,6 +1,6 @@
 from backend.modules.parsers.codeparser import CodeParser
 from backend.modules.parsers.markdownparser import MarkdownParser
-from backend.modules.parsers.MultiModalPdfParser.parser import PdfMultiModalParser
+from backend.modules.parsers.multimodal.parser import MultiModalParser
 from backend.modules.parsers.parser import register_parser
 from backend.modules.parsers.pdfparser_fast import PdfParserUsingPyMuPDF
 from backend.modules.parsers.tablepdfparser import PdfTableParser
@@ -11,5 +11,5 @@ register_parser("MarkdownParser", MarkdownParser)
 register_parser("TextParser", TextParser)
 register_parser("PdfParserFast", PdfParserUsingPyMuPDF)
 register_parser("PdfTableParser", PdfTableParser)
-register_parser("PdfMultiModal", PdfMultiModalParser)
+register_parser("MultiModalParser", MultiModalParser)
 register_parser("CodeParser", CodeParser)
