+++ backend/modules/parsers/multimodal/layout-model/mask_rcnn_X_101_32x8d_FPN_3x.yml
