+++ local.metadata.yaml
@@ -1,5 +1,5 @@
-# collection_name: bradysmall
-collection_name: brady
+collection_name: bradysmall
+# collection_name: brady
 data_source:
     type: localdir
     uri: /Users/prathamesh/Desktop/brady
@@ -7,7 +7,7 @@ parser_config:
     chunk_size: 1000
     chunk_overlap: 20
     parser_map:
-        ".pdf": PdfMultiModal
+        ".pdf": MultiModalParser
 embedder_config:
     # provider: mixedbread
     # config:
